move datatree
