+++ qiskit_ibm_runtime/quantum_program/quantum_program.py
@@ -24,7 +24,7 @@ from qiskit.circuit import QuantumCircuit
 from qiskit.quantum_info import PauliLindbladMap
 from samplomatic.samplex import Samplex
 
-from ..utils.datatree import DataTree
+from .datatree import DataTree
 
 if TYPE_CHECKING:
     from ..ibm_backend import IBMBackend
