--- qiskit_ibm_runtime/utils/datatree.py
@@ -1,24 +1,0 @@
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2025.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-"""DataTree"""
-
-from __future__ import annotations
-
-from typing import TypeAlias
-from numpy.typing import NDArray
-
-
-DataTree: TypeAlias = (
-    list["DataTree"] | dict[str, "DataTree"] | NDArray[float] | str | float | int | bool | None
-)
-"""Arbitrary nesting of lists and dicts with typed leaves."""
