+++ qiskit_ibm_runtime/utils/validations.py
@@ -146,32 +146,3 @@ def validate_job_tags(job_tags: list[str] | None) -> None:
         not isinstance(job_tags, list) or not all(isinstance(tag, str) for tag in job_tags)
     ):
         raise IBMInputValueError("job_tags needs to be a list of strings.")
-
-
-def is_datatree_compatible(data: Any) -> bool:
-    """Check if data is compatible with DataTree format.
-
-    DataTree is defined as: list["DataTree"] | dict[str, "DataTree"] | NDArray |
-    str | float | int | bool | None
-
-    Args:
-        data: The data to check.
-
-    Returns:
-        True if data is compatible with DataTree format, False otherwise.
-    """
-    if data is None or isinstance(data, (str, bool, int, float, np.ndarray)):
-        return True
-
-    if isinstance(data, list):
-        # Recursively check list elements
-        return all(is_datatree_compatible(item) for item in data)
-
-    if isinstance(data, dict):
-        # Check dict keys are strings and recursively check values
-        if not all(isinstance(key, str) for key in data.keys()):
-            return False
-        return all(is_datatree_compatible(value) for value in data.values())
-
-    # Any other type is not compatible
-    return False
